requirements.txt
exceptiongroup==1.2.0
iniconfig==2.0.0
packaging==23.2
pandoc==2.3
pluggy==1.4.0
plumbum==1.8.2
ply==3.11
pytest==8.0.0
python-slugify==8.0.3
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
